project_backend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
flake8==7.0.0
importlib_resources==6.4.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
pillow==10.3.0
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.2
zipp==3.18.1
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.2
PyJWT==2.8.0
rest-framework-simplejwt==0.0.2
